avoid define classes/objects inside of package object
scala compiler says

```
it is not recommended to define classes/objects inside of package objects.
```
