+++ atmosphere/src/main/scala/org/scalatra/atmosphere/package.scala
@@ -11,23 +11,6 @@ package object atmosphere {
 
   type AtmoReceive = PartialFunction[InboundMessage, Unit]
 
-  abstract class ClientFilter(val uuid: String) extends Function[AtmosphereResource, Boolean]
-
-  class Everyone extends ClientFilter(null) {
-    def apply(v1: AtmosphereResource): Boolean = true
-    override def toString(): String = "Everyone"
-  }
-
-  class OnlySelf(uuid: String) extends ClientFilter(uuid) {
-    def apply(v1: AtmosphereResource): Boolean = v1.uuid == uuid
-    override def toString(): String = "OnlySelf"
-  }
-
-  class SkipSelf(uuid: String) extends ClientFilter(uuid) {
-    def apply(v1: AtmosphereResource): Boolean = v1.uuid != uuid
-    override def toString(): String = "Others"
-  }
-
   val AtmosphereClientKey = "org.scalatra.atmosphere.AtmosphereClientConnection"
   val AtmosphereRouteKey = "org.scalatra.atmosphere.AtmosphereRoute"
   val ActorSystemKey = "org.scalatra.atmosphere.ActorSystem"
