+++ core/src/main/scala/org/scalatra/util/conversion/TypeConverter.scala
@@ -1,0 +1,6 @@
+package org.scalatra.util.conversion
+
+import scala.annotation.implicitNotFound
+
+@implicitNotFound(msg = "Cannot find a TypeConverter type class from ${S} to ${T}")
+trait TypeConverter[S, T] { def apply(s: S): Option[T] }
