+++ swagger/src/main/scala/org/scalatra/swagger/reflect/ParameterNameReader.scala
@@ -1,0 +1,14 @@
+package org.scalatra.swagger.reflect
+
+import java.lang.reflect.{ Constructor => JConstructor }
+import com.thoughtworks.paranamer.{ BytecodeReadingParanamer, CachingParanamer }
+
+trait ParameterNameReader {
+  def lookupParameterNames(constructor: JConstructor[_]): Seq[String]
+}
+
+object ParanamerReader extends ParameterNameReader {
+  private[this] val paranamer = new CachingParanamer(new BytecodeReadingParanamer)
+  def lookupParameterNames(constructor: JConstructor[_]): Seq[String] =
+    paranamer.lookupParameterNames(constructor)
+}
